# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FitLog API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/photo:
    post:
      operationId: uploadProfilePhoto
      tags: [profile]
      summary: Upload profile photo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        "200":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUploadResponse"
...
  /workouts:
    get:
      operationId: getWorkouts
      tags: [workouts]
      summary: Get user workouts
      parameters:
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of workouts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutListResponse"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Create a new workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutRequest"
      responses:
        "201":
          description: Workout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{id}:
    get:
      operationId: getWorkout
      tags: [workouts]
      summary: Get a specific workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Workout details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    put:
      operationId: updateWorkout
      tags: [workouts]
      summary: Update a workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutRequest"
      responses:
        "200":
          description: Workout updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    delete:
      operationId: deleteWorkout
      tags: [workouts]
      summary: Delete a workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Workout deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /measurements:
    get:
      operationId: getMeasurements
      tags: [measurements]
      summary: Get body measurements
      parameters:
        - name: days
          in: query
          schema:
            type: number
          description: Number of days to look back (30, 90, 365)
      responses:
        "200":
          description: Body measurements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeasurementListResponse"
    post:
      operationId: createMeasurement
      tags: [measurements]
      summary: Log body measurement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeasurementRequest"
      responses:
        "201":
          description: Measurement logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BodyMeasurement"
  /measurements/{id}:
    delete:
      operationId: deleteMeasurement
      tags: [measurements]
      summary: Delete a measurement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Measurement deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # PROGRESS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    PhotoUploadResponse:
      type: object
      properties:
        url:
          type: string
      required: [url]

    # AUTH SCHEMAS
...
    Profile:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        photoUrl:
          type: string
          nullable: true
        age:
          type: number
          nullable: true
        gender:
          type: string
          nullable: true
        heightCm:
          type: number
          nullable: true
        weightKg:
          type: number
          nullable: true
        fitnessGoals:
          type: array
          items:
            type: string
        activityLevel:
          type: string
          nullable: true
        dailyCalorieGoal:
          type: number
          nullable: true
        dailyProteinGoal:
          type: number
          nullable: true
        dailyCarbsGoal:
          type: number
          nullable: true
        dailyFatGoal:
          type: number
          nullable: true
      required: [id, userId, firstName, lastName, fitnessGoals]
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        age:
          type: number
          nullable: true
        gender:
          type: string
          nullable: true
        heightCm:
          type: number
          nullable: true
        weightKg:
          type: number
          nullable: true
        fitnessGoals:
          type: array
          items:
            type: string
        activityLevel:
          type: string
          nullable: true
        dailyCalorieGoal:
          type: number
          nullable: true
        dailyProteinGoal:
          type: number
          nullable: true
        dailyCarbsGoal:
          type: number
          nullable: true
        dailyFatGoal:
          type: number
          nullable: true
      required: [firstName, lastName]

    # WORKOUT SCHEMAS
    WorkoutExercise:
      type: object
      properties:
        id:
          type: number
        workoutId:
          type: number
        name:
          type: string
        order:
          type: number
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSet"
      required: [id, workoutId, name, order, sets]
    WorkoutSet:
      type: object
      properties:
        id:
          type: number
        exerciseId:
          type: number
        reps:
          type: number
          nullable: true
        weightKg:
          type: number
          nullable: true
        order:
          type: number
      required: [id, exerciseId, order]
    Workout:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        activityType:
          type: string
        name:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        durationMinutes:
          type: number
          nullable: true
        distanceKm:
          type: number
          nullable: true
        caloriesBurned:
          type: number
          nullable: true
        mood:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExercise"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, activityType, date, exercises, createdAt]
    CreateWorkoutRequest:
      type: object
      properties:
        activityType:
          type: string
        name:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        durationMinutes:
          type: number
          nullable: true
        distanceKm:
          type: number
          nullable: true
        caloriesBurned:
          type: number
          nullable: true
        mood:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/CreateExerciseRequest"
      required: [activityType, date]
    CreateExerciseRequest:
      type: object
      properties:
        name:
          type: string
        order:
          type: number
        sets:
          type: array
          items:
            $ref: "#/components/schemas/CreateSetRequest"
      required: [name, order, sets]
    CreateSetRequest:
      type: object
      properties:
        reps:
          type: number
          nullable: true
        weightKg:
          type: number
          nullable: true
        order:
          type: number
      required: [order]
    WorkoutListResponse:
      type: object
      properties:
        workouts:
          type: array
          items:
            $ref: "#/components/schemas/Workout"
        total:
          type: number
      required: [workouts, total]
...
    BodyMeasurement:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        date:
          type: string
          format: date-time
        weightKg:
          type: number
          nullable: true
        bodyFatPercent:
          type: number
          nullable: true
        chestCm:
          type: number
          nullable: true
        waistCm:
          type: number
          nullable: true
        hipsCm:
          type: number
          nullable: true
        armsCm:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, date, createdAt]
    CreateMeasurementRequest:
      type: object
      properties:
        date:
          type: string
          format: date-time
        weightKg:
          type: number
          nullable: true
        bodyFatPercent:
          type: number
          nullable: true
        chestCm:
          type: number
          nullable: true
        waistCm:
          type: number
          nullable: true
        hipsCm:
          type: number
          nullable: true
        armsCm:
          type: number
          nullable: true
      required: [date]
    MeasurementListResponse:
      type: object
      properties:
        measurements:
          type: array
          items:
            $ref: "#/components/schemas/BodyMeasurement"
      required: [measurements]

    # PROGRESS SCHEMAS
...