# lib/api-zod/src/generated/types/measurementListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FitLog API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BodyMeasurement } from "./bodyMeasurement";

export interface MeasurementListResponse {
  measurements: BodyMeasurement[];
}